+++ tests/Bridge/Pdo/PdoDriverTest.php
@@ -13,7 +13,7 @@ class PdoDriverTest extends PdoTestCase
         self::assertSame(
             1,
             (int) $this
-                ->getBridge()
+                ->getDatabaseSession()
                 ->select()
                 ->columnRaw('1', 'foo')
                 ->executeQuery()
@@ -29,7 +29,7 @@ class PdoDriverTest extends PdoTestCase
     public function testParameterTypeGuess(): void
     {
         $query = $this
-            ->getBridge()
+            ->getDatabaseSession()
             ->raw(
                 'select ?',
                 [