+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Next
+
+* [feature] `Bridge::getSchemaManager()` method is moved into the
+  `DatabaseSession` interface instead. Bridges should now never be used
+  as a public API for end users.
+* [bc] ⚠️ All errors in the `MakinaCorpus\QueryBuilder\Error\Bridge`
+  namespace are now in the `MakinaCorpus\QueryBuilder\Error\Server` namespace.
+* [bc] ⚠️ `FunctionalTestCaseTrait::getBridge()` was renamed to
+  `FunctionalTestCaseTrait::getDatabaseSession()`.
+
 ## 1.4.0
 
 * [feature] ⭐️ Added `DatabaseSession::getVendorName()`, `DatabaseSession::getVendorVersion()`,